moved functions; in progress
